+++ @ornikar/browserslist-config/degraded-support.js
@@ -2,6 +2,4 @@
 
 // Note: if you update this, update the link in README.md !
 
-module.exports = 'last 4 versions, > .2% in FR, Firefox ESR, not ie < 10, not ios < 10, not op_mini all'.split(
-  ','
-);
+module.exports = 'last 4 versions, > .2% in FR, Firefox ESR, not ie < 10, not ios < 10, not op_mini all'.split(',');