+++ @ornikar/eslint-config-babel/index.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = {
+  extends: ['@ornikar/eslint-config-babel-use'].map(require.resolve),
+
+  plugins: ['babel'],
+  parser: 'babel-eslint',
+};