+++ @ornikar/eslint-config-react/index.js
@@ -2,12 +2,15 @@
 
 module.exports = {
   extends: [
+    '@ornikar/eslint-config-babel',
     'eslint-config-airbnb/rules/react',
     'eslint-config-airbnb/rules/react-a11y',
     'eslint-config-prettier',
     'eslint-config-prettier/react',
     './rules/react',
   ].map(require.resolve),
+
+  plugins: ['eslint-plugin-react-hooks'],
 
   settings: {
     'import/resolver': {
@@ -18,7 +21,16 @@ module.exports = {
   },
 
   rules: {
-    'import/extensions': ['error', { extensions: ['js'] }],
+    'import/extensions': [
+      'error',
+      'ignorePackages',
+      {
+        js: 'never',
+      },
+    ],
     'react/jsx-filename-extension': ['error', { extensions: ['js'] }],
+
+    'react-hooks/rules-of-hooks': 'error',
+    'react-hooks/exhaustive-deps': 'error',
   },
 };