+++ @ornikar/eslint-config-typescript-react/index.js
@@ -1,0 +1,27 @@
+'use strict';
+
+module.exports = {
+  extends: [require.resolve('@ornikar/eslint-config-react')],
+
+  settings: {
+    'import/resolver': {
+      node: {
+        extensions: ['.js', '.ts', '.tsx'],
+      },
+    },
+  },
+
+  rules: {
+    'import/extensions': [
+      'error',
+      'ignorePackages',
+      {
+        js: 'never',
+        ts: 'never',
+        tsx: 'never',
+      },
+    ],
+    'react/jsx-filename-extension': ['error', { extensions: ['tsx', 'js'] }],
+    'react/prop-types': 'off',
+  },
+};