+++ @ornikar/jest-config-react/__mocks__/@storybook/react.js
@@ -3,28 +3,83 @@
 'use strict';
 
 const { shallow } = require('enzyme');
+const { shallowToJson } = require('enzyme-to-json');
+
+const decorateStory = (storyFn, decorators) =>
+  decorators.reduce(
+    (decorated, decorator) => (context = {}) =>
+      decorator(
+        (p = {}) =>
+          decorated(
+            // MUTATION !
+            Object.assign(
+              context,
+              p,
+              {
+                parameters: Object.assign(context.parameters || {}, p.parameters),
+              },
+              { options: Object.assign(context.options || {}, p.options) },
+            ),
+          ),
+        context,
+      ),
+    storyFn,
+  );
 
 // Mocked version of `import { action } from '@storybook/react'`.
-exports.action = actionName => jest.fn();
+exports.action = (actionName) => jest.fn();
 
 // Mocked version of: `import { storiesOf } from '@storybook/react'`
-exports.storiesOf = groupName => {
+exports.storiesOf = (groupName) => {
+  const localDecorators = [];
+  const localParameters = {};
+
   // Mocked API to generate tests from & snapshot stories.
   const api = {
-    add(storyName, story) {
+    add(storyName, story, storyParameters = {}) {
+      const parameters = Object.assign({ name: storyName }, localParameters, storyParameters);
+      const { jest } = parameters;
+      const { componentToTest, ignore, ignoreDecorators } = jest || {};
+
+      if (ignore) return api;
+
       describe(groupName, () => {
         it(storyName, () => {
-          expect(
-            shallow(story(), { disableLifecycleMethods: true })
-          ).toMatchSnapshot();
+          const wrappingComponent = ignoreDecorators
+            ? undefined
+            : ({ children }) => decorateStory(() => children, localDecorators)(parameters);
+
+          const wrapper = shallow(story(parameters), {
+            disableLifecycleMethods: true,
+            wrappingComponent,
+          });
+
+          if (componentToTest) {
+            const component = wrapper.find(componentToTest);
+            component.forEach((child) => {
+              expect(shallowToJson(child.dive())).toMatchSnapshot();
+            });
+          } else {
+            expect(shallowToJson(wrapper)).toMatchSnapshot();
+          }
         });
       });
 
       return api;
     },
 
+    addWithPercyOptions(storyName, percyOptions, story) {
+      api.add(storyName, story);
+    },
+
+    addParameters(parameters) {
+      Object.assign(localParameters, parameters);
+      return api;
+    },
+
     // Any `storybook-addon-*` packages may require noop-ing them:
-    addDecorator() {
+    addDecorator(decorator) {
+      localDecorators.push(decorator);
       return api;
     },
   };