+++ @ornikar/storybook-config/defaultParameters.js
@@ -1,0 +1,8 @@
+'use strict';
+
+const viewports = require('./parameters/viewports');
+
+/** default ornikar storybook parameters */
+module.exports = {
+  viewport: { viewports },
+};