+++ @ornikar/stylelint-config/index.js
@@ -1,11 +1,9 @@
 'use strict';
 
 module.exports = {
-  extends: [
-    'stylelint-config-standard',
-    'stylelint-config-css-modules',
-    'stylelint-config-prettier',
-  ].map(require.resolve),
+  extends: ['stylelint-config-standard', 'stylelint-config-css-modules', 'stylelint-config-prettier'].map(
+    require.resolve,
+  ),
 
   plugins: ['stylelint-order'].map(require.resolve),
 
@@ -14,5 +12,8 @@ module.exports = {
     'no-descending-specificity': null,
 
     'order/order': ['declarations', 'rules', 'at-rules'],
+
+    // https://stylelint.io/user-guide/rules/comment-word-blacklist/
+    'comment-word-blacklist': ['/^TODO:/', '/^FIXME:/'],
   },
 };