+++ @ornikar/jest-config/CHANGELOG.md
@@ -3,6 +3,137 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.4.0](https://github.com/ornikar/shared-configs/compare/@ornikar/jest-config@2.3.1...@ornikar/jest-config@2.4.0) (2019-10-15)
+
+
+### Features
+
+* **jest-config:** allow to add tests in config and scripts directories [no issue] ([#283](https://github.com/ornikar/shared-configs/issues/283)) ([252c7d6](https://github.com/ornikar/shared-configs/commit/252c7d6))
+
+
+
+
+
+## [2.3.1](https://github.com/ornikar/shared-configs/compare/@ornikar/jest-config@2.3.0...@ornikar/jest-config@2.3.1) (2019-08-05)
+
+
+### Bug Fixes
+
+* add module directories for lerna and module paths for non lerna projects ONK-3913 ([#248](https://github.com/ornikar/shared-configs/issues/248)) ([f51f297](https://github.com/ornikar/shared-configs/commit/f51f297))
+
+
+
+
+
+# [2.3.0](https://github.com/ornikar/shared-configs/compare/@ornikar/jest-config@2.2.0...@ornikar/jest-config@2.3.0) (2019-07-29)
+
+
+### Features
+
+* support lerna in jest configs ONK-3913 ([#237](https://github.com/ornikar/shared-configs/issues/237)) ([8d0cd90](https://github.com/ornikar/shared-configs/commit/8d0cd90))
+
+
+
+
+
+# [2.2.0](https://github.com/ornikar/shared-configs/compare/@ornikar/jest-config@2.1.0...@ornikar/jest-config@2.2.0) (2019-06-10)
+
+
+### Features
+
+* prettier config [no issue] ([#188](https://github.com/ornikar/shared-configs/issues/188)) ([e2844dd](https://github.com/ornikar/shared-configs/commit/e2844dd))
+
+
+
+
+
+# [2.1.0](https://github.com/ornikar/shared-configs/compare/@ornikar/jest-config@2.0.1...@ornikar/jest-config@2.1.0) (2019-04-30)
+
+
+### Features
+
+* allow both js and ts tests [no issue] ([#164](https://github.com/ornikar/shared-configs/issues/164)) ([99feb5a](https://github.com/ornikar/shared-configs/commit/99feb5a))
+
+
+
+
+
+## [2.0.1](https://github.com/ornikar/shared-configs/compare/@ornikar/jest-config@2.0.0...@ornikar/jest-config@2.0.1) (2019-04-29)
+
+
+### Bug Fixes
+
+* use new file name ([#159](https://github.com/ornikar/shared-configs/issues/159)) ([3a5ac62](https://github.com/ornikar/shared-configs/commit/3a5ac62))
+
+
+
+
+
+# [2.0.0](https://github.com/ornikar/shared-configs/compare/@ornikar/jest-config@1.3.1...@ornikar/jest-config@2.0.0) (2019-04-17)
+
+
+### Bug Fixes
+
+* jest 24 breaking change [no issue]([#147](https://github.com/ornikar/shared-configs/issues/147)) ([b5dcece](https://github.com/ornikar/shared-configs/commit/b5dcece))
+
+
+### Features
+
+* use setupFilesAfterEnv instead of setupTestFrameworkScriptFile [no issue] ([#145](https://github.com/ornikar/shared-configs/issues/145)) ([6cbe035](https://github.com/ornikar/shared-configs/commit/6cbe035))
+
+
+### BREAKING CHANGES
+
+* #145 introduced a breaking change of jest 24
+
+
+
+
+
+## [1.3.1](https://github.com/ornikar/shared-configs/compare/@ornikar/jest-config@1.3.0...@ornikar/jest-config@1.3.1) (2019-04-15)
+
+
+### Bug Fixes
+
+* support date parse function in jest [no issue] ([#143](https://github.com/ornikar/shared-configs/issues/143)) ([a44fe7a](https://github.com/ornikar/shared-configs/commit/a44fe7a))
+
+
+
+
+
+# [1.3.0](https://github.com/ornikar/shared-configs/compare/@ornikar/jest-config@1.2.1...@ornikar/jest-config@1.3.0) (2019-03-26)
+
+
+### Features
+
+* add timezone-mock in jest config [no issue] ([#130](https://github.com/ornikar/shared-configs/issues/130)) ([ecee74a](https://github.com/ornikar/shared-configs/commit/ecee74a))
+
+
+
+
+
+## [1.2.1](https://github.com/ornikar/shared-configs/compare/@ornikar/jest-config@1.2.0...@ornikar/jest-config@1.2.1) (2019-03-04)
+
+
+### Bug Fixes
+
+* path is not a function ([#116](https://github.com/ornikar/shared-configs/issues/116)) ([c6c5a30](https://github.com/ornikar/shared-configs/commit/c6c5a30))
+
+
+
+
+
+# [1.2.0](https://github.com/ornikar/shared-configs/compare/@ornikar/jest-config@1.1.0...@ornikar/jest-config@1.2.0) (2019-02-26)
+
+
+### Features
+
+* jest config add support typescript ([#104](https://github.com/ornikar/shared-configs/issues/104)) ([e5fa370](https://github.com/ornikar/shared-configs/commit/e5fa370))
+
+
+
+
+
 # [1.1.0](https://github.com/ornikar/shared-configs/compare/@ornikar/jest-config@1.0.0...@ornikar/jest-config@1.1.0) (2018-11-28)