+++ @ornikar/renovate-config/CHANGELOG.md
@@ -3,6 +3,199 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.13.2](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.13.1...@ornikar/renovate-config@1.13.2) (2019-10-09)
+
+
+### Bug Fixes
+
+* **renovate-config:** group react and react types packages [no issue] ([#259](https://github.com/ornikar/shared-configs/issues/259)) ([1d64ad3](https://github.com/ornikar/shared-configs/commit/1d64ad3))
+
+
+
+
+
+## [1.13.1](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.13.0...@ornikar/renovate-config@1.13.1) (2019-09-30)
+
+
+### Bug Fixes
+
+* **renovate-config:** only patch should be merged automaticly for [@types](https://github.com/types) [no issue] ([#260](https://github.com/ornikar/shared-configs/issues/260)) ([b9ab3a7](https://github.com/ornikar/shared-configs/commit/b9ab3a7))
+
+
+
+
+
+# [1.13.0](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.12.0...@ornikar/renovate-config@1.13.0) (2019-09-05)
+
+
+### Features
+
+* **renovate-config:** group ornikar components [no issue] ([#268](https://github.com/ornikar/shared-configs/issues/268)) ([052cf2e](https://github.com/ornikar/shared-configs/commit/052cf2e))
+
+
+
+
+
+# [1.12.0](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.11.0...@ornikar/renovate-config@1.12.0) (2019-08-06)
+
+
+### Features
+
+* renovate default masterIssueApproval [no issue] ([#251](https://github.com/ornikar/shared-configs/issues/251)) ([8fcb9cb](https://github.com/ornikar/shared-configs/commit/8fcb9cb))
+
+
+
+
+
+# [1.11.0](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.10.0...@ornikar/renovate-config@1.11.0) (2019-07-08)
+
+
+### Features
+
+* renovate automerge types [no issue] ([#215](https://github.com/ornikar/shared-configs/issues/215)) ([e86f8ea](https://github.com/ornikar/shared-configs/commit/e86f8ea))
+
+
+
+
+
+# [1.10.0](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.9.0...@ornikar/renovate-config@1.10.0) (2019-07-02)
+
+
+### Bug Fixes
+
+* raise renovate scheduled to every week on monday ([#208](https://github.com/ornikar/shared-configs/issues/208)) ([a5dce58](https://github.com/ornikar/shared-configs/commit/a5dce58))
+
+
+### Features
+
+* enable master issue approval for majors [no issue] ([#211](https://github.com/ornikar/shared-configs/issues/211)) ([21fddf4](https://github.com/ornikar/shared-configs/commit/21fddf4))
+
+
+
+
+
+# [1.9.0](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.8.0...@ornikar/renovate-config@1.9.0) (2019-06-10)
+
+
+### Features
+
+* add reviewflow infos in renovate prs [no issue] ([#183](https://github.com/ornikar/shared-configs/issues/183)) ([eec46fd](https://github.com/ornikar/shared-configs/commit/eec46fd))
+
+
+
+
+
+# [1.8.0](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.7.0...@ornikar/renovate-config@1.8.0) (2019-05-25)
+
+
+### Features
+
+* renovate add missing label automerge ([#186](https://github.com/ornikar/shared-configs/issues/186)) ([a84f7cc](https://github.com/ornikar/shared-configs/commit/a84f7cc))
+
+
+
+
+
+# [1.7.0](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.6.1...@ornikar/renovate-config@1.7.0) (2019-05-25)
+
+
+### Features
+
+* renovate labels config [no issue] ([#184](https://github.com/ornikar/shared-configs/issues/184)) ([5c0e4f3](https://github.com/ornikar/shared-configs/commit/5c0e4f3))
+
+
+
+
+
+## [1.6.1](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.6.0...@ornikar/renovate-config@1.6.1) (2019-05-22)
+
+**Note:** Version bump only for package @ornikar/renovate-config
+
+
+
+
+
+# [1.6.0](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.5.1...@ornikar/renovate-config@1.6.0) (2019-04-29)
+
+
+### Features
+
+* enable master issue in renovate preset ([#161](https://github.com/ornikar/shared-configs/issues/161)) ([76021c7](https://github.com/ornikar/shared-configs/commit/76021c7))
+
+
+
+
+
+## [1.5.1](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.5.0...@ornikar/renovate-config@1.5.1) (2019-04-02)
+
+
+### Bug Fixes
+
+* disable autorebase for any time scheduled prs [no issue] ([#135](https://github.com/ornikar/shared-configs/issues/135)) ([7e39f6b](https://github.com/ornikar/shared-configs/commit/7e39f6b))
+
+
+
+
+
+# [1.5.0](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.4.0...@ornikar/renovate-config@1.5.0) (2019-03-19)
+
+
+### Features
+
+* group packages renovate ONK-3275 ([#125](https://github.com/ornikar/shared-configs/issues/125)) ([44586ca](https://github.com/ornikar/shared-configs/commit/44586ca))
+* increase PR limits [no issue] ([#126](https://github.com/ornikar/shared-configs/issues/126)) ([3bbe0f1](https://github.com/ornikar/shared-configs/commit/3bbe0f1))
+* update schedule to match platform days ([#124](https://github.com/ornikar/shared-configs/issues/124)) ([ad5ee9e](https://github.com/ornikar/shared-configs/commit/ad5ee9e))
+
+
+
+
+
+# [1.4.0](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.3.1...@ornikar/renovate-config@1.4.0) (2019-03-12)
+
+
+### Features
+
+* rebase in schedule [no issue] ([#117](https://github.com/ornikar/shared-configs/issues/117)) ([8e782b4](https://github.com/ornikar/shared-configs/commit/8e782b4))
+* **renovate-config:** enable dedupe ([#118](https://github.com/ornikar/shared-configs/issues/118)) ([ddf6918](https://github.com/ornikar/shared-configs/commit/ddf6918))
+
+
+
+
+
+## [1.3.1](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.3.0...@ornikar/renovate-config@1.3.1) (2019-02-27)
+
+
+### Bug Fixes
+
+* allow rebasing in PRs ([#109](https://github.com/ornikar/shared-configs/issues/109)) ([0feb238](https://github.com/ornikar/shared-configs/commit/0feb238))
+
+
+
+
+
+# [1.3.0](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.2.1...@ornikar/renovate-config@1.3.0) (2019-02-19)
+
+
+### Features
+
+* group eslint updates [no issue] ([#101](https://github.com/ornikar/shared-configs/issues/101)) ([bc068b1](https://github.com/ornikar/shared-configs/commit/bc068b1))
+* rebase stale pr on early mondays ([#100](https://github.com/ornikar/shared-configs/issues/100)) ([02ecaa9](https://github.com/ornikar/shared-configs/commit/02ecaa9))
+
+
+
+
+
+## [1.2.1](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.2.0...@ornikar/renovate-config@1.2.1) (2019-02-19)
+
+
+### Bug Fixes
+
+* typo in package.json ([#89](https://github.com/ornikar/shared-configs/issues/89)) ([9fe6dcd](https://github.com/ornikar/shared-configs/commit/9fe6dcd))
+
+
+
+
+
 # [1.2.0](https://github.com/ornikar/shared-configs/compare/@ornikar/renovate-config@1.0.0...@ornikar/renovate-config@1.2.0) (2019-01-02)